Cleaned up defines
git-svn-id: https://svn.usrportage.de/php-mogilefs@117 6ae05860-910a-4080-baf6-befe656c970c

+++ trunk/mogilefs.c
@@ -426,7 +426,7 @@ PHPAPI int mogilefs_sock_get(zval *id, MogilefsSock **mogilefs_sock TSRMLS_DC) {
 }
 /* }}} */
 
-PHPAPI int mogilefs_sock_write(MogilefsSock *mogilefs_sock, char *cmd, int cmd_len, int free_cmd TSRMLS_DC) { /* {{{ */
+PHPAPI int mogilefs_sock_write(MogilefsSock *mogilefs_sock, char *cmd, int cmd_len, short free_cmd TSRMLS_DC) { /* {{{ */
 	int retval = 0;
 
 #ifdef MOGILEFS_DEBUG
